split PayloadWriter strategies into their own files. Renamed IsBigEndian to ShouldWriteBigEndian, to be more clear in what's going on.
